moved labeller helper class to its own directory
