+++ src/app/dataview/databar/labeller/labeller.ts
@@ -1,5 +1,5 @@
-import { DatabarComponent } from './databar.component';
-import { Label } from '../labelstreams/labelstream';
+import { DatabarComponent } from '../databar.component';
+import { Label } from '../../labelstreams/labelstream';
 import * as d3 from "d3";
 
 export class Labeller {
